write out chunks based on dependency tree
prior to this patch, this package would spit out all chunks. this worked
for webpack 3 but is incompatible with webpack 4 -- we need to actually
read the dependency tree from the webpack build artifact and spit out
only the appropriate chunks.

also add a `chunkOrder` field to the output which we can read to
determine what order bundles need to be loaded by the browser